\name{createXGMML}
\alias{createXGMML}
\title{createXGMML}
\usage{
  createXGMML(gr, node.label, cwName = "test",
    node.form = NULL, edge.form = NULL, attributes = NULL,
    path)
}
\arguments{
  \item{gr}{a graphNEL object generated by edge matrices
  provided by reac2cor.}

  \item{node.label}{a character vector containing the name
  of each node.}

  \item{cwName}{Cytoscape window name.}

  \item{node.form}{data.frame containing the names of the
  nodes in the first column, and TRUE/FALSE vectors for
  graph parameters in the additional columns. The column
  name should have the parameter type, color and shape and
  the value in the format "color.#FF0000".}

  \item{edge.form}{data.frame containing the names of the
  nodes in the first column, and TRUE/FALSE vectors for
  graph parameters in the additional columns. The column
  name should have the parameter type, color, width and
  style, and the value in the format "color.#FF0000".}

  \item{attributes}{classification table created by
  export.class.table function.}

  \item{path}{system path to save the file.}
}
\value{
  Cytoscape xgmml readable file.
}
\description{
  Creates a Cytoscape xgmml readable file.
}

