\name{design.connection}
\alias{design.connection}
\title{design.connection}
\usage{
  design.connection(reactionM)
}
\arguments{
  \item{reactionM}{a data.frame with compounds and its
  connections.}
}
\value{
  A binary matrix w of connections between candidate
  compound.
}
\description{
  Design a connection matrix w from unique reaction
  identifiers from experimentally known reactions between
  candidate compound.
}

