\name{export2cytoscape}
\alias{export2cytoscape}
\title{export2cytoscape}
\usage{
  export2cytoscape(gr, node.label, cwName = "test",
    node.form = NULL, edge.form = NULL, cpdInfo = NULL,
    classTable = NULL, pos = NULL)
}
\arguments{
  \item{gr}{a graphNEL object generated by edge matrices
  provided by reac2cor.}

  \item{node.label}{a character vector containing the name
  of each node.}

  \item{cwName}{Cytoscape window name.}

  \item{node.form}{data.frame containing the names of the
  nodes in the first column, and TRUE/FALSE vectors for
  graph parameters in the additional columns. The column
  name should have the parameter type, color and shape and
  the value in the format "color.#FF0000".}

  \item{edge.form}{data.frame containing the names of the
  nodes in the first column, and TRUE/FALSE vectors for
  graph parameters in the additional columns. The column
  name should have the parameter type, color, width and
  style, and the value in the format "color.#FF0000".}

  \item{cpdInfo}{matrix of pathway information, with
  pathway code, name and the counting how many nodes each
  pathway has.}

  \item{classTable}{matrix of classification generated by
  export.class.table. If the user provides this matrix the
  information is sent to Cytoscape as node attributes, and
  the user can visualize it in the data panel.}

  \item{pos}{node position list given by get.kgml.positions
  function.}
}
\value{
  Exports a graph to Cytoscape window, with additional
  parameters.
}
\description{
  Exports a graph to Cytoscape with optional node and edge
  attributes.
}

