\name{formula2mass}
\alias{formula2mass}
\title{formula2mass}
\usage{
  formula2mass(formula)
}
\arguments{
  \item{formula}{molecular formula, not allowing characters
  aside atom's alphabet.}
}
\value{
  a float representing the sum of monoisotopic masses.
}
\description{
  Calculates the exact mass of a given formula.
}

