\name{get.compounds.by.pathway.biocyc}
\alias{get.compounds.by.pathway.biocyc}
\title{get.compounds.by.pathway.biocyc}
\usage{
  get.compounds.by.pathway.biocyc(pathID)
}
\arguments{
  \item{pathID}{code.}
}
\value{
  character matrix of compound codes, names and formulas.
}
\description{
  Get biocyc compound codes from its API.
}

