\name{get.pathway.by.compound.kegg}
\alias{get.pathway.by.compound.kegg}
\title{get.pathway.by.compound.kegg}
\usage{
  get.pathway.by.compound.kegg(keggId)
}
\arguments{
  \item{keggId}{KEGG's compound id.}
}
\value{
  A string with the pathways collapsed by ";" character.
}
\description{
  Takes KEGG id code and retrieves the pathways that the
  compound is possible involved.
}

