\name{get.pathway.by.organism.biocyc}
\alias{get.pathway.by.organism.biocyc}
\title{get.pathway.by.organism.biocyc}
\usage{
  get.pathway.by.organism.biocyc(orgID)
}
\arguments{
  \item{orgID}{code.}
}
\value{
  character vector of pathway codes.
}
\description{
  Get biocyc pathway codes from its API.
}

