\name{get.pathway.by.organism.kegg}
\alias{get.pathway.by.organism.kegg}
\title{get.pathway.by.organism.kegg}
\usage{
  get.pathway.by.organism.kegg(organismId)
}
\arguments{
  \item{organismId}{KEGG's organism id.}
}
\value{
  A vector of KEGG pathway codes.
}
\description{
  Get KEGG pathway codes from its API.
}

