\name{get.reactions.by.compound.biocyc}
\alias{get.reactions.by.compound.biocyc}
\title{get.reactions.by.compound.biocyc}
\usage{
  get.reactions.by.compound.biocyc(cpdID)
}
\arguments{
  \item{cpdID}{code.}
}
\value{
  character vector of reaction concatenated by ";"
  character.
}
\description{
  Get biocyc reaction codes from its API.
}

