\name{KEGGcpds}
     \docType{data}
     \alias{KEGGcpds}
     \title{Compounds of KEGG database}
     \description{
	This is a illustrative database format for ProbMetab
	package, with the mandatory fields id, formula and
	reactions.       

     }
     \usage{KEGGcpds}
     \format{A data.fram containing 14014 compound entries.}
     \source{http://www.kegg.jp/.}
     \references{
	Kanehisa, Minoru and Goto, Susumu (2000) KEGG: kyoto encyclopedia of genes and genomes
	Nucleic acids research. 28(1):27-30.
     }
     \keyword{datasets}
