\name{NIST_relative_atomic_mass}
     \docType{data}
     \alias{NIST_relative_atomic_mass}
     \title{Most abundant atomic mass From NIST database.}
     \description{
	This is a periodic table used by ProbMetab::formula2mass 
	to calculate the exact mass of a given molecular formula.
     }
     \usage{NIST_relative_atomic_mass}
     \format{A data.fram containing 118 atomic masses.}
     \source{http://www.nist.gov/pml/data/comp.cfm.}
     \references{
The National Institute of Standards and Technology (NIST) is an agency of the U.S. Department of Commerce.
     }
     \keyword{datasets}
