\name{openGraph}
\alias{openGraph}
\title{openGraph}
\usage{
  openGraph(JSONFile, classTable = NULL,
    openBrowser = FALSE)
}
\arguments{
  \item{JSONFile}{a file generated by createJSONToCytoscape
  function.}

  \item{classTable}{classification table generated by
  export.class.table function.}

  \item{openBrowser}{logical if FALSE, returns a URL link,
  if TRUE opens the web browser to access graph
  visualization.}
}
\description{
  Send the JSONFile to http://labpib.fmrp.usp.br/ web
  server and to visualize a dynamic graph in the web
  browser.
}

