\name{RCreateDataFrame}
\alias{RCreateDataFrame}
\title{RCreateDataFrame}
\usage{
  RCreateDataFrame(file)
}
\arguments{
  \item{file}{xml input file.}
}
\value{
  a data frame with compound's id, name and formula.
}
\description{
  Create a data frame with compound information from
  mzMatch open source project format.
}

