\name{weightM}
\alias{weightM}
\title{weightM}
\usage{
  weightM(isoPatt, useIso = TRUE, intervals = NULL,
    offset = NULL, massWeigth = 1, likelihood = "erfc",
    precision = 1)
}
\arguments{
  \item{isoPatt}{is the likelihood data.frame generated by
  incorporate.isotopes function, or reactionM the
  data.frame of compound's information.}

  \item{useIso}{logical indicating whether to use or not
  isotope information in the likelihood.}

  \item{intervals}{a vector of SNR numerical intervals, to
  which different carbon offset should be added to
  predicted C-number.}

  \item{offset}{vector of empirically estimated carbon
  offset to be added to predicted C-number.}

  \item{massWeigth}{is the contribution parameter of the
  probabilistic model.}

  \item{likelihood}{which noise model to use, "erfc" to
  complementary error function, or "gaussian" to standard
  gaussian with two sd corresponding to the given p.p.m
  precision.}

  \item{precision}{equipment mass accuracy, usually the
  same used in exact mass search.}
}
\value{
  A matrix wm of likelihood weights.
}
\description{
  Builds a c (number of compounds) by m (number of masses)
  matrix of compound likelihoods with the gaussian error
  function erf, based on mass accuracy and optionally on
  isotopic patterns.
}

